Terraformer v2
Uploading the entire eclipse project - hope that helps! If you want to use another IDE - the only file you need outside of src/ is lib/antlr-4.8-complete.jar.

As for the breakdown of the contents, I'll explain it to you later. It's still a work in progress.

The functionality in high priority:
1. tf2json
2. json2tf
3. tf2tokens